feat: add platform connectivity service
