+++ pubspec.yaml
@@ -56,10 +56,11 @@ dependencies:
   
   # UUID Generation
   uuid: ^4.2.1
-  
+
   # File Operations
   path_provider: ^2.1.1
   share_plus: ^7.2.1
+  connectivity_plus: ^5.0.2
 
 dev_dependencies:
   flutter_test:
